Added -Newfile switch to Set-LogMode
I had a use case where an infinite loop needed to trigger a script on an interval and create a log file each time. Due to Plog's caching behavior, each run of the script was saving to the same log file.

This exposes a parameter to allow me as a user to force creation of a new log file.